Move getFileFromHTTP to xmltv.py to avoid library conflict
- move function
- remove imports
- add an extra check for older bottle versions
