+++ lang/tvstreamrecord.russian.json
@@ -114,7 +114,6 @@
     "From .m3u file": "Из .m3u файла",
     "From HTTP URL": "Из HTTP URL",
 
-
     "Electronic Program Guide": "Электронная телепрограмма",
     "Current date": "Текущая дата",
     "Keyword": "Ключевое слово",
